+++ payments/RELEASES.md
@@ -28,3 +28,13 @@ Based on:
 - [typescript v1.0.1] payments
 ### Releases
 - [NPM v1.0.1] https://www.npmjs.com/package/@wingspan/payments/v/1.0.1 - payments
+
+## 2024-01-22 01:30:52
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.148.0 (2.237.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.0] payments
+### Releases
+- [NPM v1.1.0] https://www.npmjs.com/package/@wingspan/payments/v/1.1.0 - payments